# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerStatus": {
        "enum": [
          "answered",
          "abstained"
        ],
        "title": "AnswerStatus",
        "type": "string"
      },
...
      "ConversationCreate": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ConversationCreate",
        "type": "object"
      },
      "ConversationResponse": {
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_query_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Query Id"
          },
          "latest_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Question"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "title",
          "turn_count",
          "latest_query_id",
          "latest_question",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryConceptContext": {
        "additionalProperties": false,
        "properties": {
          "anchors": {
            "items": {
              "$ref": "#/components/schemas/QueryConceptAnchor"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Anchors",
            "type": "array"
          },
          "label": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "anchors"
        ],
        "title": "QueryConceptContext",
        "type": "object"
      },
      "QueryCreate": {
        "additionalProperties": false,
        "properties": {
          "concept_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryConceptContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "conversation_id": {
            "anyOf": [
              {
                "maxLength": 36,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          },
          "question": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryCreate",
        "type": "object"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuredAnswer"
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "query_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Intent"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "retrieval_diagnostic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Diagnostic"
          },
          "retrieval_rounds": {
            "items": {
              "$ref": "#/components/schemas/RetrievalRoundResponse"
            },
            "title": "Retrieval Rounds",
            "type": "array"
          },
          "standalone_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standalone Question"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trace": {
            "$ref": "#/components/schemas/QueryTraceResponse"
          },
          "usage": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Usage",
            "type": "object"
          }
        },
        "required": [
          "id",
          "course_id",
          "conversation_id",
          "question",
          "status",
          "answer",
          "failure_code",
          "usage",
          "query_intent",
          "standalone_question",
          "retrieval_rounds",
          "retrieval_diagnostic",
          "trace",
          "created_at",
          "completed_at"
        ],
        "title": "QueryResponse",
        "type": "object"
      },
      "QueryTraceResponse": {
        "properties": {
          "retrieval_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Snapshot Id"
          },
          "retrieval_trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Trace Id"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "retrieval_snapshot_id",
          "retrieval_trace_id"
        ],
        "title": "QueryTraceResponse",
        "type": "object"
      },
...
      "RetrievalRoundResponse": {
        "properties": {
          "active_index": {
            "title": "Active Index",
            "type": "boolean"
          },
          "active_lexical_index_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Lexical Index Id"
          },
          "candidate_count": {
            "title": "Candidate Count",
            "type": "integer"
          },
          "eligible_count": {
            "title": "Eligible Count",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "retrieval_trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Trace Id"
          }
        },
        "required": [
          "query",
          "active_index",
          "candidate_count",
          "eligible_count",
          "retrieval_trace_id",
          "active_lexical_index_id"
        ],
        "title": "RetrievalRoundResponse",
        "type": "object"
      },
...
      "StructuredAnswer": {
        "additionalProperties": false,
        "properties": {
          "answer_markdown": {
            "title": "Answer Markdown",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "title": "Claims",
            "type": "array"
          },
          "query_id": {
            "minLength": 1,
            "title": "Query Id",
            "type": "string"
          },
          "refusal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Refusal"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnswerStatus"
          }
        },
        "required": [
          "query_id",
          "status",
          "answer_markdown"
        ],
        "title": "StructuredAnswer",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "期末复习智能体 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/conversations/{conversation_id}/queries": {
      "get": {
        "operationId": "list_conversation_queries_api_v1_conversations__conversation_id__queries_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueryResponse"
                  },
                  "title": "Response List Conversation Queries Api V1 Conversations  Conversation Id  Queries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Conversation Queries",
        "tags": [
          "queries"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/conversations": {
      "get": {
        "operationId": "list_conversations_api_v1_courses__course_id__conversations_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationResponse"
                  },
                  "title": "Response List Conversations Api V1 Courses  Course Id  Conversations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Conversations",
        "tags": [
          "queries"
        ]
      },
      "post": {
        "operationId": "create_conversation_api_v1_courses__course_id__conversations_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Conversation",
        "tags": [
          "queries"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/queries": {
      "get": {
        "operationId": "list_queries_api_v1_courses__course_id__queries_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueryResponse"
                  },
                  "title": "Response List Queries Api V1 Courses  Course Id  Queries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Queries",
        "tags": [
          "queries"
        ]
      },
      "post": {
        "operationId": "create_query_api_v1_courses__course_id__queries_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Query",
        "tags": [
          "queries"
        ]
      }
    },
...
    "/api/v1/queries/{query_id}": {
      "get": {
        "operationId": "get_query_api_v1_queries__query_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "title": "Query Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Query",
        "tags": [
          "queries"
        ]
      }
    },
...
    "/api/v1/queries/{query_id}/events": {
      "get": {
        "operationId": "stream_query_events_api_v1_queries__query_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "title": "Query Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "once",
            "required": false,
            "schema": {
              "default": false,
              "title": "Once",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Query Events",
        "tags": [
          "queries"
        ]
      }
    },
...
  }
}